requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1

gitpython==3.1.43
chardet==5.2.0
pathspec==0.12.1

sentence-transformers==2.7.0
chromadb==0.5.3

ollama==0.3.3